refactor: remove test_support modules in favor of direct re-exports
