+++ docs/development-guide.md
@@ -13,10 +13,6 @@
 The workspace follows the original template's testing culture:
 
 - Unit tests live next to owner modules.
-- Internal test doubles are owner-local and exported through:
-	- `src/snippets/test_support.rs`
-	- `src/context_files/test_support.rs`
-	- `src/clipboard/test_support.rs`
-	- `src/project_fs/test_support.rs`
+- Internal test doubles are owner-local and exported directly through their respective `mod.rs` (`src/snippets/mod.rs`, `src/context_files/mod.rs`, `src/clipboard/mod.rs`, `src/project_fs/mod.rs`).
 - Integration tests are organized by concern at `tests/cli/`, `tests/context/`, `tests/snippets/`, and `tests/security/`.
 - Shared integration fixtures are centralized under `tests/harness/`.
