+++ src/clipboard/mod.rs
@@ -5,7 +5,7 @@ mod system_clipboard;
 #[cfg(test)]
 mod in_memory_clipboard;
 #[cfg(test)]
-pub mod test_support;
+pub use in_memory_clipboard::InMemoryClipboard;
 
 use crate::error::AppError;
 use std::env;
