--- src/clipboard/test_support.rs
@@ -1,1 +1,0 @@
-pub use super::in_memory_clipboard::InMemoryClipboard;
