+++ src/context_files/mod.rs
@@ -6,7 +6,7 @@ mod local_context_store;
 #[cfg(test)]
 mod in_memory_context_store;
 #[cfg(test)]
-pub mod test_support;
+pub use in_memory_context_store::InMemoryContextStore;
 
 pub use context_file_store::{ContextFileStore, ContextWriteStatus};
 pub use key::{resolve_context_path, resolve_validated_context_path};
