--- src/project_fs/test_support.rs
@@ -1,2 +1,0 @@
-pub use super::fixed_workspace_locator::FixedWorkspaceLocator;
-pub use super::in_memory_workspace_file_reader::InMemoryWorkspaceFileReader;
