--- src/snippets/test_support.rs
@@ -1,3 +1,0 @@
-pub use super::in_memory_catalog::InMemoryCatalog;
-pub use super::in_memory_checkout::InMemoryCheckout;
-pub use super::in_memory_snippet_store::InMemorySnippetStore;
